+++ Aelia's Apprentice/Assets/Scripts/Player.cs
@@ -7,6 +7,8 @@ public class Player : MonoBehaviour
     [SerializeField] private GameObject spellPrefab3;
     [SerializeField] private GameObject spellPrefab4;
 
+    SpellQueue spellQueue;
+
     private GameObject InteractorBox { get; set; }
 
     void Start()
@@ -22,11 +24,43 @@ public class Player : MonoBehaviour
 
         if (spellPrefab4 == null)
             Debug.LogError(gameObject + " needs to have a spell prefab 4");
+
+        spellQueue = new SpellQueue();
     }
 
     void Update()
     {
         if (Input.GetKeyDown(KeyCode.Alpha1))
+        {
+            spellQueue.AddSpell(SpellWord.Fire);
+        }
+
+        if (Input.GetKeyDown(KeyCode.Alpha2))
+        {
+            spellQueue.AddSpell(SpellWord.Push);
+        }
+
+        if (Input.GetKeyDown(KeyCode.Alpha3))
+        {
+            spellQueue.AddSpell(SpellWord.Size);
+        }
+
+        if (Input.GetKeyDown(KeyCode.Alpha4))
+        {
+            spellQueue.AddSpell(SpellWord.Reverse);
+        }
+
+        if (Input.GetKeyDown(KeyCode.Space))
+        {
+            CastSpell();
+            spellQueue.ClearQueue();
+        }
+    }
+
+    void CastSpell()
+    {
+        // If spell has fire but no ice, create a damage projectile
+        if (spellQueue.Contains(SpellWord.Fire) && !spellQueue.Contains(SpellWord.Reverse))
         {
             Vector2 spawnPos = transform.position;
             Vector2 direction = Vector2.up;
@@ -35,17 +69,25 @@ public class Player : MonoBehaviour
 
             GameObject projectileObj = Instantiate(spellPrefab1, spawnPos, rotation);
             projectileObj.GetComponent<Projectile>().Direction = direction;
+            
+            // Exit early since spell has been cast
+            return;
         }
 
-        if (Input.GetKeyDown(KeyCode.Alpha2))
+        // If spell contains fire and reverse, create a freeze area
+        if (spellQueue.Contains(SpellWord.Fire))
         {
             Vector2 spawnPos = transform.position;
             Quaternion rotation = Quaternion.identity;
 
             GameObject areaObj = Instantiate(spellPrefab2, spawnPos, rotation);
+
+            // Exit early since spell has been cast
+            return;
         }
 
-        if (Input.GetKeyDown(KeyCode.Alpha3))
+        // If spell contains push and no fire, create a push projectile 
+        if (spellQueue.Contains(SpellWord.Push))
         {
             Vector2 spawnPos = transform.position;
             Vector2 direction = Vector2.up;
@@ -54,9 +96,13 @@ public class Player : MonoBehaviour
 
             GameObject projectileObj = Instantiate(spellPrefab3, spawnPos, rotation);
             projectileObj.GetComponent<Projectile>().Direction = direction;
+
+            // Exit early since spell has been cast
+            return;
         }
 
-        if (Input.GetKeyDown(KeyCode.Alpha4))
+        // If spell has size and no fire or push, cast a size projectile
+        if (spellQueue.Contains(SpellWord.Size))
         {
             Vector2 spawnPos = transform.position;
             Vector2 direction = Vector2.up;
@@ -65,6 +111,9 @@ public class Player : MonoBehaviour
 
             GameObject projectileObj = Instantiate(spellPrefab4, spawnPos, rotation);
             projectileObj.GetComponent<Projectile>().Direction = direction;
+
+            // Exit early since spell has been cast
+            return;
         }
     }
 }